+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
+## 2022 - 2022/01/23
+
+- Bump pillow version (CVE-2022-22815, CVE-2022-22816, CVE-2022-22817)
+- Update deps
+- Minor update to tests
+
 ## 2021.1.4 - 2021/12/10
 
 - Raise `NotImplementedError` for XcfDocument save methods https://github.com/FHPythonUtils/GimpFormats/pull/10